Split WysiwygUITests
