+++ platforms/ios/example/WysiwygUITests/WysiwygUITests.swift
@@ -17,7 +17,7 @@
 import XCTest
 
 class WysiwygUITests: XCTestCase {
-    private let app = XCUIApplication(bundleIdentifier: "org.matrix.Wysiwyg")
+    internal let app = XCUIApplication(bundleIdentifier: "org.matrix.Wysiwyg")
 
     override func setUpWithError() throws {
         continueAfterFailure = false
@@ -26,112 +26,6 @@ class WysiwygUITests: XCTestCase {
     }
 
     override func tearDownWithError() throws { }
-
-    /// Type a text and delete some different kind of text selections with the composer.
-    func testTypingAndDeleting() throws {
-        // Type something into composer.
-        textView.typeTextCharByChar("abc🎉🎉👩🏿‍🚀")
-        assertTextViewContent("abc🎉🎉👩🏿‍🚀")
-
-        // Test deleting parts of the text.
-        let deleteKey = app.keys["delete"]
-        deleteKey.tap()
-        assertTextViewContent("abc🎉🎉")
-
-        let delete3CharString = String(repeating: XCUIKeyboardKey.delete.rawValue, count: 3)
-        textView.typeTextCharByChar(delete3CharString)
-        assertTextViewContent("ab")
-
-        // Rewrite some content.
-        textView.typeTextCharByChar("cde 🥳 fgh")
-        assertTextViewContent("abcde 🥳 fgh")
-
-        // Double tap results in selecting the last word.
-        textView.doubleTap()
-        deleteKey.tap()
-        // Note: iOS is removing the whitespace right after the emoji, even though it reports
-        // through `shouldChangeTextIn` that it is removing only the 3 last chars.
-        assertTextViewContent("abcde 🥳")
-
-        // Triple tap selects the entire line.
-        textView.tap(withNumberOfTaps: 3, numberOfTouches: 1)
-        deleteKey.tap()
-        assertTextViewContent("")
-    }
-
-    /// Type a text and make it bold in the composer.
-    /// A screenshot is saved since string attributes can't be read from this context.
-    func testTypingAndBolding() throws {
-        // Type something into composer.
-        textView.typeTextCharByChar("Some bold text")
-
-        textView.doubleTap()
-        // We can't detect data being properly reported back to the model but
-        // 1s is more than enough for the Rust side to get notified for the selection.
-        sleep(1)
-
-        button(.boldButton).tap()
-        // Bolding doesn't change text and we can't test text attributes from this context.
-        assertTextViewContent("Some bold text")
-
-        // Keep a screenshot of the bolded text.
-        let screenshot = textView.screenshot()
-        let attachment = XCTAttachment(screenshot: screenshot)
-        attachment.lifetime = .keepAlways
-        add(attachment)
-    }
-
-    /// Type and send a message with the composer.
-    ///
-    /// Expected plain text content is "Some bold text" and
-    /// HTML representation is "Some bold <strong>text</strong>"
-    func testTypingAndSending() throws {
-        // Type something into composer.
-        textView.typeTextCharByChar("Some bold text")
-
-        textView.doubleTap()
-        // 1s is more than enough for the Rust side to get notified for the selection.
-        sleep(1)
-        button(.boldButton).tap()
-        // We can't detect data being properly reported back to the model but
-        // 1s is more than enough for the Rust side to get notified for the selection.
-        sleep(1)
-        button(.sendButton).tap()
-
-        XCTAssertEqual(staticText(.contentText).label, "Some bold __text__")
-        XCTAssertEqual(staticText(.htmlContentText).label, "Some bold <strong>text</strong>")
-    }
-    
-    // Remember to disable hardware keyboard and use only software keyboard for this UITest
-    func testTypingFast() throws {
-        let text = "Some long text that I am going to type very fast"
-        textView.tap()
-        sleep(1)
-        textView.typeText(text)
-        let options = XCTExpectedFailure.Options()
-        options.isStrict = false
-        XCTExpectFailure("Typing fast might fail on CI", options: options)
-        assertTextViewContent(text)
-    }
-    
-    func testLongPressDelete() throws {
-        let multilineText =
-            """
-            test1
-            test2
-            test3
-            test4
-            test5
-            test6
-            test7
-            test8
-            test9
-            test10
-            """
-        app.typeTextCharByChar(multilineText)
-        XCUIApplication().keys["delete"].press(forDuration: 15.0)
-        assertTextViewContent("")
-    }
 
     func testMinMaxResizing() throws {
         sleep(1)
@@ -143,84 +37,9 @@ class WysiwygUITests: XCTestCase {
         sleep(1)
         XCTAssertEqual(textView.frame.height, WysiwygSharedConstants.composerMinHeight)
     }
-    
-    func testCreateLinkWithTextEditAndRemove() {
-        // Create with text
-        button(.linkButton).tap()
-        XCTAssertTrue(textField(.linkUrlTextField).exists)
-        XCTAssertTrue(textField(.linkTextTextField).exists)
-        textField(.linkUrlTextField).typeTextCharByChar("url")
-        textField(.linkTextTextField).tap()
-        textField(.linkTextTextField).typeTextCharByChar("text")
-        app.buttons["Ok"].tap()
-        assertTreeEquals(
-            """
-            └>a "url"
-              └>"text"
-            """
-        )
-        
-        // Edit
-        button(.linkButton).tap()
-        XCTAssertFalse(textField(.linkTextTextField).exists)
-        textField(.linkUrlTextField).doubleTap()
-        textField(.linkUrlTextField).typeTextCharByChar("new_url")
-        app.buttons["Ok"].tap()
-        assertTreeEquals(
-            """
-            └>a "new_url"
-              └>"text"
-            """
-        )
-        
-        // Remove
-        button(.linkButton).tap()
-        XCTAssertFalse(textField(.linkTextTextField).exists)
-        app.buttons["Remove"].tap()
-        assertTreeEquals(
-            """
-            └>"text"
-            """
-        )
-    }
-    
-    func testCreateLinkFromSelection() {
-        textView.typeTextCharByChar("text")
-        assertTreeEquals(
-            """
-            └>"text"
-            """
-        )
-        
-        textView.doubleTap()
-        button(.linkButton).tap()
-        XCTAssertFalse(textField(.linkTextTextField).exists)
-        textField(.linkUrlTextField).typeTextCharByChar("url")
-        app.buttons["Ok"].tap()
-        assertTreeEquals(
-            """
-            └>a "url"
-              └>"text"
-            """
-        )
-    }
-    
-    func testTypingInlineCodeDisablesOtherFormatters() {
-        button(.inlineCodeButton).tap()
-        textView.typeTextCharByChar("code")
-        let reactiveButtonsIdentifiers: [WysiwygSharedAccessibilityIdentifier] = [
-            .boldButton,
-            .italicButton,
-            .strikeThroughButton,
-            .linkButton,
-        ]
-        for identifier in reactiveButtonsIdentifiers {
-            XCTAssertFalse(button(identifier).isEnabled)
-        }
-    }
 }
 
-private extension WysiwygUITests {
+internal extension WysiwygUITests {
     /// Returns the text view component of the composer.
     var textView: XCUIElement {
         app.textViews[rawIdentifier(.composerTextView)]
@@ -282,7 +101,7 @@ private extension WysiwygUITests {
     }
 }
 
-private extension XCUIElement {
+internal extension XCUIElement {
     /// Types a text inside the UI element character by character.
     /// This is especially useful to avoid missing some characters on
     /// UI tests running on a rather slow CI.
